Fix tuple encoders
Tuple encoders pass common tuple schema to tuple members encoders, it's
ok when members are primitive types because their encoders don't use
schema to encode value, String fallbacks to Utf8. For example if we use
Option as member of tuple it requires UNION schema to encode but gets RECORD.

To fix this every member of tuple should take only it's part of schema.