ISSUE-387 | Fixed LocalTime Encoder to properly convert to milliseconds (#388)
* ISSUE-387 | Fixed LocalTime Encoder to properly convert to milliseconds
Fixed the Encoder for LocalTime objects to properly convert from nanoseconds to millisseconds. Nano second values are truncated to the millisecond precision. LocalTime.now() produces millisecond precision values by default so the schema and encoder/decoder are inline with default functionality of the LocalTime object even though LocalTime does allow for a higher level of precision.

* ISSUE-387 | Change Schema for LocalTime to be logical type time-micros
Changed the schema for LocalTime to be logical type time-micros. In Java8 the default functionality of LocalTime.now() is to return millisecond precision, but in Java11 the default functionality is to return micro second precision. I decided to make the default schema, encoder, and decoder for LocalTime compatible with highest level of default funcitonality for now().