recursively resolve default values
if a default value is a case class with default values, they were not
correctly handled.

Use the default resolver back to resolve the default value in the
default value recursively

Known issue: at some point, we need the type name if the schema is a
union. The name used here does't resolve the correct name if an
@AvroName annotation is used on the class. Same applies on the field
name, no FieldMapper is applied.

Signed-off-by: François LAROCHE <fl@make.org>